split pregel tests
